refactor endpoint result builder
